refactor: Consolidate data preprocessing modes under `preprocess_mode` by removing the `use_patch_extraction` flag and cast `wscale` to float32.

+++ src/bacterial_gan/models/stylegan2_ada.py
@@ -70,7 +70,7 @@ class ModulatedConv2D(layers.Layer):
         in_channels = feature_shape[-1]
 
         fan_in = in_channels * self.kernel_size * self.kernel_size
-        self.wscale = self.gain / np.sqrt(fan_in)
+        self.wscale = tf.cast(self.gain / np.sqrt(fan_in), tf.float32)
 
         self.weight = self.add_weight(
             name="weight",
